fix memstats
